--- src/commands/environment/environment.ts
@@ -1,24 +1,0 @@
-/*
- * Copyright (C) 2018 Garden Technologies, Inc. <info@garden.io>
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-import { Command } from "../base"
-import { EnvironmentConfigureCommand } from "./configure"
-import { EnvironmentDestroyCommand } from "./destroy"
-
-export class EnvironmentCommand extends Command {
-  name = "environment"
-  alias = "env"
-  help = "Manage your runtime environment(s)"
-
-  subCommands = [
-    EnvironmentConfigureCommand,
-    EnvironmentDestroyCommand,
-  ]
-
-  async action() { return {} }
-}
