+++ test/src/commands/set.ts
@@ -1,11 +1,11 @@
 import { expect } from "chai"
-import { ConfigSetCommand } from "../../../../src/commands/config/set"
-import { expectError, makeTestContextA } from "../../../helpers"
+import { SetConfigCommand } from "../../../src/commands/set"
+import { expectError, makeTestContextA } from "../../helpers"
 
-describe("ConfigSetCommand", () => {
+describe("SetConfigCommand", () => {
   it("should set a config variable", async () => {
     const ctx = await makeTestContextA()
-    const command = new ConfigSetCommand()
+    const command = new SetConfigCommand()
 
     await command.action(ctx, { key: "project.mykey", value: "myvalue" })
 
@@ -14,7 +14,7 @@ describe("ConfigSetCommand", () => {
 
   it("should throw on invalid key", async () => {
     const ctx = await makeTestContextA()
-    const command = new ConfigSetCommand()
+    const command = new SetConfigCommand()
 
     await expectError(
       async () => await command.action(ctx, { key: "bla.mykey", value: "ble" }),
